Simplified food node modules
